Created a new View for the order selection.
Created a new OrderSelection View and Controller to be able to select an order for the burger on assembly.
Modified DrawingManager to enable the drawing of an Hamburger where you want.
Added methods in DrawingManager to generate and draw an order.
Added a orderNumber variable and a copyOf() method in OrderImpl.
Added a variable selectedOrder in GameModel.
Also fixed some small issues.